move cheatcode tests to separate contract
