+++ src/tests/integration/data/contracts/test-cheatcodes/.gitignore
@@ -1,0 +1,2 @@
+/target
+.env
