+++ src/tests/integration/data/contracts/test-cheatcodes/rust-toolchain.toml
@@ -1,0 +1,2 @@
+[toolchain]
+channel = "1.83.0"
