+++ src/tests/integration/data/contracts/test-hello-world/src/lib.rs
@@ -1,8 +1,10 @@
 #![cfg_attr(not(any(test, feature = "export-abi")), no_main)]
 extern crate alloc;
 
-use alloy_primitives::address;
-use stylus_sdk::{alloy_primitives::U256, prelude::*};
+use stylus_sdk::{
+    alloy_primitives::{address, FixedBytes, U256},
+    prelude::*,
+};
 
 use skribe::{build_init_code, cheat};
 
@@ -14,7 +16,6 @@ sol_interface! {
   }
 }
 
-// `Counter` will be the entrypoint.
 sol_storage! {
     #[entrypoint]
     pub struct TestCounter {
@@ -71,25 +72,4 @@ impl TestCounter {
         assert_eq!(a, b);
     }
 
-    pub fn test_deal(&mut self) {
-        let alice = address!("AABBCCDDEEFF0011223344556677889900AABBCC");
-        assert_eq!(self.vm().balance(alice), U256::ZERO);
-
-        let balance = U256::from(123000000000u128);
-        cheat().deal(&mut *self, alice, balance).unwrap();
-
-        assert_eq!(self.vm().balance(alice), balance);
-    }
-
-    pub fn test_env_setters(&mut self) {
-        // warp
-        let time = 123000000000u64;
-        cheat().warp(&mut *self, U256::from(time)).unwrap();
-        assert_eq!(self.vm().block_timestamp(), time);
-        
-        // roll
-        let block_number = 123;
-        cheat().roll(&mut *self, U256::from(block_number)).unwrap();
-        assert_eq!(self.vm().block_number(), block_number);
-    }
 }
